requirements.txt
fastapi
uvicorn
requests
python-docx
PyMuPDF
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-groq
langchain_together
python-dotenv
faiss-cpu
pypdf
python-docx
unstructured